ml-service/requirements.txt
fastapi
uvicorn
motor
pymongo
pandas
scikit-learn
numpy
python-dotenv
apscheduler
